+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-05-22 20:48:57
-// Commit: test(state): implement widget integration tests for state
+// Updated at: 2026-05-23 17:25:56
+// Commit: style(state): harmonize light and dark theme palette for state
 
 class stateComponentState {
-  final String commitRef = '1336';
-  final String lastMessage = 'test(state): implement widget integration tests for state';
-  final DateTime timestamp = DateTime.parse('2026-05-22T20:48:57+05:00');
+  final String commitRef = '1355';
+  final String lastMessage = 'style(state): harmonize light and dark theme palette for state';
+  final DateTime timestamp = DateTime.parse('2026-05-23T17:25:56+05:00');
 }